refactor FraHome components
