+++ webapp/pages/AssessmentHome/FraHome/SelectedCountries/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SelectedCountries'
